Refactor input and output of wrapper
The input configuration of the UIPS wrapper depended on reading
text-input files. This has now been changed to use a new class
UIPSOptions that encapsulates all these parameters. Additionally,
instead of writing the results of the downsampling to file, these
are now returned as part of the function.

To demonstrate the behavior of this new interface a jupyter
notebook based on the 2D Gaussian example was added to the
tutorials repo.

Additional work remains to support the previous interfaces as well